+++ clash-vhdl/CHANGELOG.md
@@ -3,7 +3,7 @@
 ## 0.6.4
 * New features:
   * Support for Haskell's: `Char`, `Int8`, `Int16`, `Int32`, `Int64`, `Word`, `Word8`, `Word16`, `Word32`, `Word64`.
-  * Int/Word/Integer bitwidth for generated VHDL is configurable using the `-clash-intwidth=N` flag, where `N` is the number of bits.
+  * Int/Word/Integer bitwidth for generated VHDL is configurable using the `-clash-intwidth=N` flag, where `N` can be either 32 or 64.
 
 ## 0.6.4 *November 17th 2015*
 * Fixes bugs: